# api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: JIRA Tool API
  version: 1.0.0
  description: API for a JIRA Tool
paths:
...
  /api/v1/projects/{projectId}/tasks/{taskId}/attachments/:
    get:
      tags:
        - attachments
      summary: Get a file attachment
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
        - name: attachmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: File attachment details
    post:
      tags:
        - attachments
      summary: Upload a file attachment
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: File uploaded
...